feat : add services folder and openAI.ts
